pyproject.toml
[project]
name = "webdoze"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black[d]>=25.1.0",
    "python-fasthtml>=0.12.19",
]
